QA/telegram/telegram.py: increase robustness to flaky results
Sometimes a single telegram DC fails. So, we need to figure out a more
robust way of checking individual results. Until that point limit ourselves
to just checking top-level keys. This means we can significantly reduce
the timeout between each run, which was bumped in the hope that it could
help to reduce the frequency of flaky DC results.

Part of https://github.com/ooni/probe-engine/issues/356