webconnectivity: fix the semantics of blocking
According to the implementations, the semantics is that blocking is false
when accessible is true, a nullable string otherwise.

The diff documents in detail how and why this was introduced.

It would probably have been better to not make this field semantics so
complicate, but now it's too late. Maybe in the future we can have some
extra field with a more clear and simple semantics?

Part of https://github.com/ooni/probe-engine/issues/810