QA/webconnectivity.py: add test case for nonexistent domain (#842)
The behavior of probe-engine is better than the one of MK. We have
disabled this check for MK and we should open an issue.

The following diff shows the difference in the toplevel keys:

```diff
--- MK.jsonl	2020-08-13 13:12:05.741791884 +0200
+++ OPE.jsonl	2020-08-13 13:12:14.092840549 +0200
@@ -1,9 +1,9 @@
 {
   "input": "http://antani.xyz",
   "test_keys": {
-    "accessible": true,
+    "accessible": null,
     "agent": "redirect",
-    "blocking": false,
+    "blocking": null,
     "body_length_match": null,
     "body_proportion": 0,
     "client_resolver": "91.80.36.88",
@@ -16,15 +16,16 @@
         "body_length": -1,
         "failure": "dns_lookup_error",
         "headers": {},
-        "status_code": -1
+        "status_code": -1,
+        "title": ""
       },
       "tcp_connect": {}
     },
     "control_failure": null,
     "dns_consistency": "consistent",
-    "dns_experiment_failure": null,
+    "dns_experiment_failure": "dns_nxdomain_error",
     "headers_match": null,
-    "http_experiment_failure": "dns_lookup_error",
+    "http_experiment_failure": null,
     "retries": null,
     "socksproxy": null,
     "status_code_match": null,
```

Because `antani.xyz` does not exist, OPE is clearly right here. I am also
wondering whether this issue with MK is related to the following issue:

https://github.com/ooni/probe-engine/issues/579

Maybe?^H^H^H Most likely

What I find rather interesting, also, is that there is no indication of
"dns_experiment_failure" in MK even though it clearly failed. Yet, we see
consistency in the "dns_consistency" key. But, still, I would argue it's
not so smart to elide the DNS failure in this case.

Also, note that the HTTP experiment is not even performed by OPE because
we really don't know what to do without IP addresses for the domain.