QA/webconnectivity: blocked IP w/ consistent DNS
This is a rather common blocking case. Let us make sure we
measure this case with OPE like we do with MK.

I needed to remove the checks pertaining to the in depth
structure of the JSON, because fields are missing in JSON
files generated by MK. We know these bugs.

Part of https://github.com/ooni/probe-engine/issues/810