QA/webconnectivity: TCP/IP blocking w/ inconsistent DNS
We open a listening port on localhost but we don't listen to such
port. We redirect HTTP to such port. So connect fails.

We also hijack the DNS such that the target host is localhost.

While there, fix printing the DNS consistency check result.

I have manually checked we match MK's behaviour.