QA/Dockerfile: pin to container using Go 1.14
Fixes https://github.com/ooni/probe-engine/pull/854/checks?check_run_id=997008056#step:4:426

The problem seems to be that in Go 1.15 the qtls.ConnectionState is not
compatible with the tls.ConnectionState.

Because we're using Alpine Edge, this means the build that failed was
using Go 1.15, while the previous builds used Go 1.14.

To fix the build, let us pin to a Docker container using Go 1.14.