QA/webconnectivity: introduce status flags
The status flags allow us to measure in a more precise way why and how
we flag a measurement as such. The flags are currently marked as an
experimental feature. I see them as a tool to:

1. make sure we can assert more precisely in tests why a specific
censorship condition led to a specific result

2. see quite clearly what are the places in which our understanding
of what went wrong is not very precise

After this last round of QA, I have covered most of the error conditions
specified in summary.go, and verified that the behavior that we are
implementing now is consistent with MK in such cases.

Therefore, with the merging of this diff, we will be reaching the point
where https://github.com/ooni/probe-engine/issues/852 is done.