Vendor gocertifi and force cURL's certificates
I noticed that certifi/gocertifi was not updated for quite some time.

While there, I also noticed they are not using the CA we were using with
MK, even though it's probably generated from the same source.

All of this led me to the decision of vendoring the script used to
generate the certificate chain, and vendor cURL's CA.

Part of https://github.com/ooni/probe-engine/issues/748.