urlgetter: set timeout when it's used directly
When urlgetter is used as a library, the code using urlgetter shall
configure the proper timeout that makes sense for the use case.

When we're calling urlgetter directly from the command line, instead, we
now set a conservative, large enough 45 seconds timeout.

This will allow users running urlgetter directly to receive timeouts
if the code is stuck fetching for any reason.

Because urlgetter is not meant to fetch large files directly anyway, this
number seems reasonable. I could have made the timeout configurable from
the command line, and I actually originally did that. Yet, I noticed it was
complicating the code and especially reasoning about which was the proper
piece of code responsibile of setting the timeout.

So, I decided to give priority to code simplicity in light of the fact
that in the end urlgetter is mainly a library.

Part of https://github.com/ooni/probe-engine/issues/655.