Improve documentation and consistency (2/n) (#877)
* resources: this should be a top-level package

Fetching resources is integral to OONI operations. We want this
functionality to be a full class citizen rather than hidden.

Again, part of https://github.com/ooni/probe-engine/issues/746
for the blog post on the new engine in Go.

* httpfailure: avoid too many internal packages put in top-level one

Again for the blog post https://github.com/ooni/probe-engine/issues/746

* version: this doesn't need to be a separate package

All the usages of version.Version are in the engine package or below.

Therefore, it's fine to merge into engine.

Again, for the blog post https://github.com/ooni/probe-engine/issues/746.