[API CHANGE] libminiooni: give up backwards compatibility and document
This cannot be a fast moving research client if we are strongly constrained
by backwards compatibility with previous CLI interfaces.

Also, I discovered that, sadly, we've already broken OONI Probe v2.x CLI
by giving another semantics to `-i` and `-O`.

Therefore, really be explicit in our goal of moving fast and do research and
hence give up on backwards compatibility for miniooni.

While there, yet, aim for some consistency and use `--probe-services` as
a command line flag rather than `--probe-services-url`, which is redundant.

Also, remove unused `--bouncer` and `--collector` options which were only
cluttering the output and making the code more complicated.

Do not remove the `--no-geoip` option because it seems we're going to have to
support this functionality again in the future, to deal with the use case in
which the internet is mostly shutdown and we still wanna measure.

Part of https://github.com/ooni/probe-engine/issues/746