libooniffi: add docs and basic smoke test (#887)
This package is not used in any OONI product, so we may break it without noticing. Users depending on it should be aware of that, as I have abundantly made clear in the documentation. Nonetheless, I added a very simple smoke test, such that, at least, we notice if we break the build. I considered deleting this package, but it seems useful to have, as long as people using it knows that this is not something we officially support, so we don't pay much attention to it. What we actually use is `oonimkall`, and this package is just a tiny wrapper around it. For this reason, I eventually chose to keep it with a little more tests.

Part of general revamp of documentation, tests, and architecture for preparing https://github.com/ooni/probe-engine/issues/746.