github/workflows: simulate a user integrating probe-engine
We want to have some confidence that a user can setup a Go repository
and use probe-engine as a library. While this seems simple, our current
dependency structure is complicated by the fact that psiphon does not
have a go.mod file, so sometimes there are breakages.

Part of https://github.com/ooni/probe-engine/issues/746