oonimkall: implement SubmitTask
With the availability of the SubmitTask, we now have code to which
we can pass measurements to resubmit without being concerned of whether
they belong to the same report or not. The code will indeed to the
right thing and we will open/close reports as needed.

Design at https://github.com/ooni/probe-engine/issues/893#issuecomment-686613123