DESIGN.md: restrict the scope of the design document
We've added new APIs in https://github.com/ooni/probe-engine/issues/893
but we didn't feel the need of updating DESIGN.md, rather it was more
productive to use an issue as a designing tool.

This is all fine and we should do whatever works(tm) but we should also
make sure we don't provide false expectations with DESIGN.md.

A cheap way to do that is to clearly mark this design document scope
to be that of describing how to replace Measurement Kit.