cmd/webscore: rescore previous web measurements with the Go engine
This diff introduces a new PoC where we score previous measurements using
the scoring mechanism in the OONI engine. This should in principle allow us
to evaluate whether we're able to detect and workaround old false positive
conditions. The main trouble is that past measurements in many cases have
less fields than now. So, what I have done in this diff is to just recompute
the overall result w/o recomputing keys like `dns_consistent`.

This exercise has been useful to bootstrap an investigation on whether we
could make sense of a bunch of anomalies, as documented by

https://github.com/ooni/probe-engine/issues/770

More works needs to be done to continue this research in future Sprints.