libminiooni: allow newline at EOF with -f/--file
It didn't occur to me initially, but the fact that vim adds a newline at
the end of the file adds an extra blank input to be tested.

We actually don't want to emit an error for such an input and the fix
is just to avoid adding blank inputs to the inputs list.

Since the fix is so easy, while there allow any other blank line in
the middle of the file, since it costs us nothing.

Part of https://github.com/ooni/probe-engine/issues/926