fix(model/experiment.go): start splitting ExperimentSession
ExperimentSession is a large interface in the model package that is
too large and too widely used to be easily modifiable.

I noticed this issue when I started investigating the issue at
https://github.com/ooni/explorer/issues/495.

Before we embark into more substantial changes, let us clear the
table a little bit by reducing the scope of ExperimentSession.