webconnectivity: we can now safely omit client resolver IP
Since this experiment has been rewritten in Go, we can safely omit the
IP address of the client resolver. The new implementation, in fact,
performs a geolookup of such an IP address.

Not strictly and directly related with the issue I am working on
(https://github.com/ooni/explorer/issues/495). Yet, working on such
an issue and reading the code is a useful mindset to reason about
simplifications around what we submit to our backends.