session: add privacy preserving geolocation accessors (#964)
The previous diff (https://github.com/ooni/probe-engine/pull/963) has
introduced the concept of an accessor for the resolver IP whose returned
value changes depending on the configured privacy settings.

This is a good programming practice. So, while I am slowly approaching to
the bottom of https://github.com/ooni/explorer/issues/495, I though it was
good yak shaving to implement accessors for every variable for which the
configured privacy settings matter.