fix(experiment.go): always use privacy preserving geolocation accessors
With 7ffe9dc49b98e23253a3571d330410e300a8ce0f, we introduced privacy preserving
geolocation accessors. With this commit, we start using them.

It turns out the original problem that triggered this investigation (i.e.,
https://github.com/ooni/explorer/issues/495) happens because the code inside
experiment.go sends the real ASN when opening the report. This is not the
correct behavior, which is to access the ASN through the the privacy
preserving wrapper. Such that, if the user does not want to share their
ASN with OONI, the report is opened using "AS0".

Upon inspection, it seems we need to apply the same fix also to other fields
of the openReport and Measurement structures. And, even if we were not needing
that, it seems the most correct thing to do anyway.

This diff does precisely that. It ensures we always use the new accessors
and adds a bunch of new tests to ensure this is WAI.