fix(model): make sure we honour sharing settings (#968)
With https://github.com/ooni/probe-engine/pull/966 and this change
together we have a more in depth mechanism for enforcing that we're
doing the right thing in terms of honouring sharing settings.

Yet, my instinct is that we should push for removing this vestigial
options that increase complexity and reduce the quality of data we
receive, in the mid term. Shall discuss this with the rest of the team.