doc: improve routine sprint releases
We are a bit far from this objective, but ideally we would like a release in probe-engine to cause downstream releases into probe-cli (already happening now), probe-android and probe-ios (not happening now), probe-desktop (not happening now). The way in which this process should work, in my view, is that these downstream releases should be development releases. Ideally we would like a bunch of people to be using routinely using development releases (which we kinda obtain once we have better support for running tests automatically). Ideally we would like to be able to promote a specific build from unstable to stable without rebuilding.

Because this change only addresses documentation, there is no point in creating a PR. I am going to commit directly to the master branch.