fix: review when and how frequently we run tests
This is a combination of 5 commits:

1. fix: run the generate check every night at midnight

This check is very lightweight and it makes sense to run it during
the night such that it can inform our next day development.  A breakage
in this check on day X can be fixed on day X+1.

Also this commit is part of https://github.com/ooni/probe-engine/issues/1005
where we start by asking the question whether we're running too many tests
for each commit. As part of this review, I have actually spooted this specific
test which instead runs too infrequently and could be improved.

2. fix: run go1.15 checks every day at midnight

We're investigating test breakages as part of issue
https://github.com/ooni/probe-engine/issues/1005.

While there, we're reviewing all the tests we run and
making sure they run frequently enough.

This test in particular can also be run every day
at midnight, so we can fix a failure happening
on a specific day on the following day.

3. fix: run libooniffi, jafar integration tests at night

These integration tests are fine to run during the night, since we
do not modify jafar and libooniffi very often. However, let's run these
during the night rather than in the middle of the work day.

Part of https://github.com/ooni/probe-engine/issues/1005.

4. fix: run miniooni checks every night at midnight

Part of https://github.com/ooni/probe-engine/issues/1005. This test
wasn't running frequently enough. Every day is better.

5. fix: run the using test weekly during the night

Part of the review of when and why we run tests initiated when
looking at https://github.com/ooni/probe-engine/issues/1005.

This is clearly fine to run weekly but we're happier with
running this experiment at night rather than in the middle
of the work day (for people working from Europe).