Start addressing #1005's CI failures (#1007)
* fix: run QA checks daily

The issue described at https://github.com/ooni/probe-engine/issues/1005
is caused by many underlying issues in our codebase.

One such issue is that we're running to many tests for every commit.

This makes development slower than it could be. We should ideally run
quick tests for every commit and daily comprehensive tests.

This commit moves us in the right direction by making sure that QA
checks run nightly (for me) and reasonably spaced in time.

* fix: review when and how frequently we run tests

This is a combination of 5 commits:

1. fix: run the generate check every night at midnight

This check is very lightweight and it makes sense to run it during
the night such that it can inform our next day development.  A breakage
in this check on day X can be fixed on day X+1.

Also this commit is part of https://github.com/ooni/probe-engine/issues/1005
where we start by asking the question whether we're running too many tests
for each commit. As part of this review, I have actually spooted this specific
test which instead runs too infrequently and could be improved.

2. fix: run go1.15 checks every day at midnight

We're investigating test breakages as part of issue
https://github.com/ooni/probe-engine/issues/1005.

While there, we're reviewing all the tests we run and
making sure they run frequently enough.

This test in particular can also be run every day
at midnight, so we can fix a failure happening
on a specific day on the following day.

3. fix: run libooniffi, jafar integration tests at night

These integration tests are fine to run during the night, since we
do not modify jafar and libooniffi very often. However, let's run these
during the night rather than in the middle of the work day.

Part of https://github.com/ooni/probe-engine/issues/1005.

4. fix: run miniooni checks every night at midnight

Part of https://github.com/ooni/probe-engine/issues/1005. This test
wasn't running frequently enough. Every day is better.

5. fix: run the using test weekly during the night

Part of the review of when and why we run tests initiated when
looking at https://github.com/ooni/probe-engine/issues/1005.

This is clearly fine to run weekly but we're happier with
running this experiment at night rather than in the middle
of the work day (for people working from Europe).

* fix: only run golang tests on pull request

All other checks should run periodically but they should
otherwise not run for pushes and pull requests.

Part of the effort to simplify and reduce the number of checks
run for every commit that aims to eventually solve the issue
described at https://github.com/ooni/probe-engine/issues/1005.

* fix: only use powerdns inside of sessionresolver

Looking at the failure described by https://github.com/ooni/probe-engine/issues/1005,
it seems we're over relying on powerdns in our tests.

So let's try and not use powerdns except for sessionresolver.

* fix: move failing test in the integration tests category

Related to https://github.com/ooni/probe-engine/issues/1005. We are
addressing the problem as follows. Let us start separating integration
testing from unit testing (for now in a corse grained way).

We know that the tests that are failing are all integration tests.
So, by making integration tests optional and by not running them
we return to a green situation. Of course we will pay a big price
in terms of code coverage. But this also tells us that we were
obtaining coverage in the wrong way.

Once the master branch is green again, and with a reduced coverage,
we can spend some more time curating the tests. I would like to
see tests run on every commit of a PR to converge within a small
amount of minutes. Maybe the integration flag is not the right
flag, but let us obtain a green tree, then we can improve.

* fix: remove -v from go test invocation

Should produce a more manageable output. Still doing this in the
context of https://github.com/ooni/probe-engine/issues/1005.