fix: only run short tests (#1008)
* fix: only run short tests

For pull requests (and possibly also for master) we generally want to run
short tests because this allows for faster feedback cycles.

To this end, this PR strives to flag as "long" tests that take too much time
and also cuts the time consumed by tests that can be easily optimised.

Part of https://github.com/ooni/probe-engine/issues/1005.

* fix: print coverage of each package

This makes the test runs output more useful. This work is still
part of https://github.com/ooni/probe-engine/issues/1005.