fix(all): remove the integration tag
As I noticed in https://github.com/ooni/probe-engine/issues/1005#issuecomment-719276445,
this tag is actually harmful because it prevents VSCode from taking into account
the specific file, so a bunch of workflows are more complex.

Also, regardless of VSCode, if you don't build with the right tag you cannot notice
cases where you have, e.g., conflicting declarations. So, it's much better to
have all tests building by default and to explicitly skip long ones. Even though
this entails some manual work from me.

In the meanwhile, update CONTRIBUTING.md.

Also, remove the convention where we have `TestUnit` and `TestIntegration`
sometimes, because it was not being followed consistently. When this action
caused a test to be named just `Test`, I've changed it to `TestGood`.

There will be some extra work to do, to maybe reorganise some large
`foo_integration_test.go` files that now also contain simple, very quick,
but currently skipped tests, but this is trivial and for later.

(With VSCode able to act on files, it's actually quite easy to figure
out which are the fast tests to run at every commit.)