fix: allow building w/o -tags PSIPHON_DISABLE_QUIC (#1022)
Releasing a version of probe-engine that only builds with `-tags PSIPHON_DISABLE_QUIC` for every version of Go is a bad idea. This PR is based of https://github.com/ooni/psiphon, which is vendors psiphon sources and all its dependencies in a way that does not cause conflicts in the package path.

This codebase also disables QUIC support when using Go 1.15 (now the default on Fedora), so we don't have any conflict and the code builds seamlessly. Because I've been told by Psiphon developers that they're working to upgrade their QUIC implementation, this means we are not going to use Psiphon without QUIC for a long time when using Go 1.15. Also, they told me it should ~fine to use Psiphon without QUIC anyway. Lastly, we're currently using Go 1.14 for building our own packages, so we're not removing anything from our official packages.

With this diff committed, we dramatically simplify dependency management and we remove the requirement of specifying a manual flag to get probe-engine (and hence probe-cli) to compile with Go 1.15.

This is the reference issue: https://github.com/ooni/probe-engine/issues/866.