feature: run more integration tests on release branches
We have an action item saying that we should ensure all tests are
green when approaching a release. To facilitate this job, let us
enable more integration tests when approaching a release. In this
commit, I'm adding the ones that make more sense.

This work is done as part of https://github.com/ooni/probe-engine/issues/1004