fix(urlgetter): do not reuse session's tunnel
My diagnosis of https://github.com/ooni/probe-cli/issues/173 is that
psiphon causes a psiphon tunnel to start and to be installed into the
session causing riseupvpn to subsequently use such a tunnel.

So, for example:

```bash
./miniooni --tunnel=psiphon riseupvpn
```

Uses psiphon for _both_ speaking with the probe services and speaking
with the riseupvpn backend, which is not WAI.

The underlying problem seems to be that an experiment can mutate a
session. I introduced this behavior for the sake of performance
experiments, but running them over psiphon and tor is not something
we are currently using.

This diff is my first attempt at addressing the original problem
by teaching urlgetter to use a private tunnel rather than changing
the behaviour of the session.

Because we probably don't care much about re-using the session tunnel
when running performance (as stated above, this is not even a use
case), I think it may be a good idea to zap this code.

I also think at this stage we're at 30% of fixing the issue because
there is need to look at tests and ensure we have no regressions.

Reference issues:

- https://github.com/ooni/probe-engine/issues/1050
- https://github.com/ooni/probe-cli/issues/173

This diff shows that we have fixed:

```diff
--- master.txt	2020-11-18 12:02:15.000000000 +0100
+++ withfix.txt	2020-11-18 12:02:59.000000000 +0100
@@ -1,5 +1,6 @@
 {
   "annotations": {
+    "_probe_engine_sanitize_test_keys": "true",
     "assets_version": "20201112191431",
     "engine_name": "ooniprobe-engine",
     "engine_version": "0.21.0-alpha",
@@ -14,12 +15,12 @@
     "tunnel": 0
   },
   "input": null,
-  "measurement_start_time": "2020-11-18 11:01:03",
+  "measurement_start_time": "2020-11-18 11:00:09",
   "probe_asn": "AS30722",
   "probe_cc": "IT",
   "probe_ip": "127.0.0.1",
   "probe_network_name": "Vodafone Italia S.p.A.",
-  "report_id": "20201118T110103Z_riseupvpn_IT_30722_n1_aL9EpK3p8OEp8NpG",
+  "report_id": "20201118T110009Z_riseupvpn_IT_30722_n1_EQrmydeuRexQL24i",
   "resolver_asn": "AS30722",
   "resolver_ip": "127.0.0.2",
   "resolver_network_name": "Vodafone Italia S.p.A.",
@@ -33,1116 +34,704 @@
       {
         "failure": null,
         "operation": "http_transaction_start",
-        "t": 9.7665e-05
+        "t": 5.217e-05
       },
       {
         "failure": null,
         "operation": "http_request_metadata",
-        "t": 0.000235983
+        "t": 0.000129534
       },
       {
-        "address": "127.0.0.1:54935",
         "failure": null,
-        "operation": "connect",
-        "proto": "tcp",
-        "t": 0.001088266
-      },
-      {
-        "failure": null,
-        "num_bytes": 3,
-        "operation": "write",
-        "t": 0.001163441
+        "operation": "resolve_start",
+        "t": 0.000203696
       },
       {
         "failure": null,
-        "num_bytes": 2,
-        "operation": "read",
-        "t": 0.001278896
-      },
-      {
-        "failure": null,
-        "num_bytes": 23,
-        "operation": "write",
-        "t": 0.00130992
+        "operation": "resolve_done",
+        "t": 0.318086661
       },
       {
+        "address": "198.252.153.107:443",
         "failure": null,
-        "num_bytes": 4,
-        "operation": "read",
-        "t": 0.587879075
-      },
-      {
-        "failure": null,
-        "num_bytes": 6,
-        "operation": "read",
-        "t": 0.587923159
+        "operation": "connect",
+        "proto": "tcp",
+        "t": 0.506691587
       },
       {
         "failure": null,
         "operation": "tls_handshake_start",
-        "t": 0.587966727
+        "t": 0.506712213
       },
       {
         "failure": null,
         "num_bytes": 286,
         "operation": "write",
-        "t": 0.588377587
+        "t": 0.506953816
       },
       {
         "failure": null,
         "num_bytes": 517,
         "operation": "read",
-        "t": 0.78868145
```

We have, in fact, domain name resolutions and connects to non localhost
hosts, meaning we're not using the tunnel now.