fix(oonimkall): adapt to removed privacy settings
We've removed the privacy settings. So adapt the tests that are
now failing. As in other parts of this PR, I've left some rather
large TODOs around. I feel like we first need to apply the core
change and then we need to deal with the fallout.

See https://github.com/ooni/probe-engine/issues/974.