feat(dash): introduce APIs used by probe-cli
Here the idea is to make the dash experiment such that we can
easily ask it about summary test keys and a log summary from
probe-cli. This allows to get rid of some code that should not
actually be part of probe-cli. In turn, the objective here is
to further the integration between the two code bases to ensure
we can tell probe-cli to run nettests using specific inputs as
documented by https://github.com/ooni/probe/issues/1283.

I'm proceeding piecemeal, adding functions to each experiment
and then ensuring all experiments have the same functions. Thus
this is just the first ~small diff of many.