fix context timeout not working as intended bug
This diff fixes the context timeout not working as intended bug
described in https://github.com/ooni/probe-engine/issues/994.

Before canceling the context we need to increase the counter so that
we're sure that canceling the context happens _after_ we increased
the counter. Otherwise, it may actually happen before depending on the
choices made by the Go scheduler.

Closes https://github.com/ooni/probe-engine/issues/994. If I am
wrong we will reopen, but I think this is at least a good explanation.