fix(oohelperd_test.go): no need to connect
If we wanna also connect we probably need to ensure we're trying to
connect only _after_ we're listening. So we need to further complicate
the code or live with a race condition.

How about doing nothing then?

This seems nice and also the reason why we're testing this main is
just so that it does not distract us when we're asking the question
of what parts of the codebase we're not actually testing.