feat: web connectivity test helper in go (#1078)
* feat: web connectivity test helper in go

This diff implements a web connectivity test helper client and
server written in go and based on the netx engine.

As part of this diff, I found convenient moving the version code
into a subpackage to significantly shrink the helper size. In fact,
the helper does not need to link with, e.g., psiphon.

The code we're introducing here works but we have not spent much
time yet checking whether it works exactly like the previous
implementation of the web connectivity test helper.

We know there is a bunch of places where it does not make much sense
to be compatible. For example, the new implementation emits netx
errors, which are more precise than the previous ones. For this reason,
it may be a good idea to just deploy the new helper and eventually
point the new probes to it. They are going to benefit a lot from this
because they could more precisely compare errors.

I have also implemented a client for the test helper because it seems
very handy to compare the new and the legacy test helpers.

The new helper is a web service because this simplifies testing and
comparing the new and the old code.

The new helper does not bother itself with caching. It's expected that
we will deal with caching most likely inside the API.

We expect, in going forward, that the test helper will evolve to provide
the API with basically the same info that a probe would provide.

Nonetheless, for this first iteration, the simplest approach of emitting
the same data format of the legacy helper has been chosen.

I have aimed for ~100% code coverage and reasonable smoke testing of
this new code, because this seems a reliable basis for evolving it in
the near future. We should nonetheless add more tests concerning the
way in which the helper behaves in case of errors.

The reference issue is https://github.com/ooni/backend/issues/276.

* fix(oohelper): better docstring

* fix(oohelperd_test.go): no need to connect

If we wanna also connect we probably need to ensure we're trying to
connect only _after_ we're listening. So we need to further complicate
the code or live with a race condition.

How about doing nothing then?

This seems nice and also the reason why we're testing this main is
just so that it does not distract us when we're asking the question
of what parts of the codebase we're not actually testing.

* fix: avoid data race