feature: adapt inputloader from libminiooni
This diff takes the no-so-tested code for loading input from
libminiooni and makes it better and much more tested.

As a side effect of this diff, miniooni now takes in input
not just a single file but a number of files.

Also, since this diff, -i and -f are not mutually exclusive,
which looked like an unnecessary restiction.

This work helps with https://github.com/ooni/probe/issues/1283.