fix(engine): InputLoader's now an interface (#1093)
We need InputLoader to be an interface because:

1. its implementation is actually quite appropriate for an interface
since we select different code paths depending on the policy

2. an interface is easier to wrap for probe-cli

While there, don't name config structs like NewFooConfig because the
NewSomething pattern in Go implies a factory, so better to avoid being
confusing with naming.

Notwithstanding point 1. above, I'm not doing a full refactoring to
an interfaces based implementation for InputLoader because the existing
code is already good and we don't need this change now.

This diff is part of https://github.com/ooni/probe/issues/1283.