cleanup: experiments should not use session tunnel
We had the possibility of using session tunnel for experiments and
the experiments using it were urlgetter, ndt7, and dash.

We have modified urlgetter to use its own tunnel. We probably also need
to ensure urlgetter and the session use different directories.

We are not using tunnels for dash and ndt7. While it would be nice to
keep this feature in tree, it complicates the tree.

So, let us remove this feature. We can always revert this diff if in
the future we discover that we need this.